simplify pairing process
