+++ drivers/p1/pair/pair.html
@@ -1,62 +1,61 @@
 <!doctype html>
 <html>
 <head>
-	<link rel="stylesheet" type="text/css" href="styles.css" />
-	<script type="text/javascript" src="/manager/webserver/assets/js/jquery.js"></script>
-	<script type="text/javascript" src="pair.js" data-origin="pair"></script>
-	<style>
-		/* place on top of Homey navigation window */
-		#hy-views { position: static; }
-	</style>
+    <script type="text/javascript">
+      Homey.setTitle(__('pair.title'));
+    </script>
+    <script src="pair.js"></script>
+    <link rel="stylesheet" type="text/css" href="styles.css"/>
 </head>
 <body>
-	<div class="setting">
-		<p data-i18n="pair.instructions" class="instructions">Instructions</p>
+<div class="setting">
+    <p data-i18n="pair.instructions" class="instructions">Instructions</p>
 
-		<label class="settings-item-description" data-i18n="pair.name">Device name</label>
-		<div class="settings-itemgroup">
-			<input id="deviceName" type="text" data-i18n="title" value="P1 smartmeter">
-		</div>
+    <label class="settings-item-description" data-i18n="pair.name">Device name</label>
+    <div class="settings-itemgroup">
+        <input id="deviceName" type="text" data-i18n="title" value="P1 smartmeter">
+    </div>
 
-		<label class="settings-item-description" data-i18n="pair.includeMeters">Meters</label>
-			<div class="settings-itemgroup">
-					<div class="settings-item-text">
-							<span data-i18n="pair.includeProduction">Include power production</span>
-					</div>
-					<div class="settings-item-last">
-							<label class="form-switch">
-									<input type="checkbox" id="includeProduction" checked="true">
-									<i></i>
-							</label>
-					</div>
-			</div>
+    <label class="settings-item-description" data-i18n="pair.includeMeters">Meters</label>
+    <div class="settings-itemgroup">
+        <div class="settings-item-text">
+            <span data-i18n="pair.includeProduction">Include power production</span>
+        </div>
+        <div class="settings-item-last">
+            <label class="form-switch">
+                <input type="checkbox" id="includeProduction" checked="true">
+                <i></i>
+            </label>
+        </div>
+    </div>
 
-			<div class="settings-itemgroup">
-				<div class="settings-item-text">
-						<span data-i18n="pair.includeOffPeak">Include power offPeak</span>
-				</div>
-				<div class="settings-item-last">
-						<label class="form-switch">
-								<input type="checkbox" id="includeOffPeak" checked="true">
-								<i></i>
-						</label>
-				</div>
-			</div>
+    <div class="settings-itemgroup">
+        <div class="settings-item-text">
+            <span data-i18n="pair.includeOffPeak">Include power offPeak</span>
+        </div>
+        <div class="settings-item-last">
+            <label class="form-switch">
+                <input type="checkbox" id="includeOffPeak" checked="true">
+                <i></i>
+            </label>
+        </div>
+    </div>
 
-			<div class="settings-itemgroup">
-					<div class="settings-item-text">
-							<span data-i18n="pair.includeGas">Include gas</span>
-					</div>
-					<div class="settings-item-last">
-							<label class="form-switch">
-									<input type="checkbox" id="includeGas" checked="true">
-									<i></i>
-							</label>
-					</div>
-			</div>
+    <div class="settings-itemgroup">
+        <div class="settings-item-text">
+            <span data-i18n="pair.includeGas">Include gas</span>
+        </div>
+        <div class="settings-item-last">
+            <label class="form-switch">
+                <input type="checkbox" id="includeGas" checked="true">
+                <i></i>
+            </label>
+        </div>
+    </div>
 
-		<label class="settings-item-description" data-i18n="pair.applySettings">Add device</label>
-		<button id="applySettings" class="btn" onclick="applySettings()" data-i18n="pair.applySettings">Add device</button>
-	</div>
+    <button id="applySettings" class="btn" onclick="pair()" data-i18n="pair.applySettings">
+        <span data-i18n="pair.addDevice">Add device</span>
+    </button>
+</div>
 </body>
 </html>
