+++ locales/nl.json
@@ -1,13 +1,14 @@
 {
     "pair": {
 		"title": "P1 slimme meter",
-		"instructions": "Deze driver kan gebruikt worden als je een P1 meter verbonden hebt door middel van com.p1.smartmeter-dsmr.reader software. De-selecteer de meters die je niet nodig hebt. De hoofd-stroommeter wordt altijd weergegeven.",
+		"instructions": "Deze driver kan gebruikt worden als je een P1 meter verbonden hebt door middel van com.p1.smartmeter-dsmr.reader software. Selecteer de meters die je wilt toevoegen. De hoofd-stroommeter wordt altijd weergegeven.",
 		"includeMeters": "Meters weergeven",
 		"includeProduction": "Stroom teruglevering",
 		"includeOffPeak": "Dal tarief",
 		"includeGas": "Gas",
 		"applySettings": "Opslaan",
         "success": "Succes!",
-        "required": "Voer een naam in voor het device!"
-    }
+        "required": "Voer een naam in voor het device!",
+		"addDevice": "Apparaat toevoegen"
+	}
 }
